-Добавил новые классы для хранения данных через Room -Исправил модель json-ответа ретрофит

+++ app/src/main/java/com/example/weatherapp/data/weather/remote/model/Astro.kt
@@ -10,8 +10,8 @@ data class Astro (
   @SerializedName("moonrise"          ) var moonrise         : String? = null,
   @SerializedName("moonset"           ) var moonset          : String? = null,
   @SerializedName("moon_phase"        ) var moonPhase        : String? = null,
-  @SerializedName("moon_illumination" ) var moonIllumination : Int?    = null,
-  @SerializedName("is_moon_up"        ) var isMoonUp         : Int?    = null,
-  @SerializedName("is_sun_up"         ) var isSunUp          : Int?    = null
+  @SerializedName("moon_illumination" ) var moonIllumination : Double?    = null,
+  @SerializedName("is_moon_up"        ) var isMoonUp         : Long?    = null,
+  @SerializedName("is_sun_up"         ) var isSunUp          : Long?    = null
 
 )
