+++ app/src/main/java/com/example/weatherapp/data/weather/remote/model/Forecastday.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Forecastday (
 
   @SerializedName("date"       ) var date      : String?         = null,
-  @SerializedName("date_epoch" ) var dateEpoch : Int?            = null,
+  @SerializedName("date_epoch" ) var dateEpoch : Long?            = null,
   @SerializedName("day"        ) var day       : Day?            = Day(),
   @SerializedName("astro"      ) var astro     : Astro?          = Astro(),
   @SerializedName("hour"       ) var hour      : ArrayList<Hour> = arrayListOf()
