+++ app/src/main/java/com/example/weatherapp/data/weather/remote/model/Location.kt
@@ -11,7 +11,7 @@ data class Location (
   @SerializedName("lat"             ) var lat            : Double? = null,
   @SerializedName("lon"             ) var lon            : Double? = null,
   @SerializedName("tz_id"           ) var tzId           : String? = null,
-  @SerializedName("localtime_epoch" ) var localtimeEpoch : Int?    = null,
+  @SerializedName("localtime_epoch" ) var localtimeEpoch : Long?    = null,
   @SerializedName("localtime"       ) var localtime      : String? = null
 
 )
